[TCFMM-7] Add auto-generated typescript API
Problem: We will eventually need to develop a small frontend to be
able to interact with the contract. To do that, however, we'll first
need to develop an low-ish level typescript API for that, generated
from the (Haskell) types.

Solution: Add typescript sdk for SegCFMM, add makefile command, update
CI.