[TCFMM-8] Add `x80` newtype
Problem: sqrt_price is kept multiplied by `2^80` over the code, however
this fact is mostly not reflected in variables so we just have to keep
that in mind, and this hardens understanding and increases the
probability of mistakes.

Solution: add a "newtype" for keeping such a value - natural fraction
with 2^-80 precision.

In some places in code they used `90` power instead of `80`.
The article says to use `80`, but there is also `TODO` in code to use
`90` for better precision. We are planning to think about `90` precision
later, and for now want to make the code consistent and correct.