[Chore] Add `.tz` file to haskell deps
Problem: currently we do not state in the Haskell's build system that
there are `.tz` files we want to depend on, we only call
`qAddDependentFile`. This way the file is indeed fetched, however once
the project is built, `cabal`/`stack` do not know that the module that
fetches the contract has to be rebuild when `.tz` file changes.

Solution: add the file to `extra-source-files` section of
`package.yaml`.

Also, move `README.md` to a separate `extra-doc-files` section instead.