[TCFMM-8] Use `X` newtype for fees growth
Problem: one another field that is actually fractional but only its
mantissa is carried - fees growth. We use `128` power of two for it.

Solution: use `X 128 Natural` there instead of plain `Natural`.

An interesting thing here is that article also mentions a related
`secondsPerLiquidityOutsideX128` field that should be carried in the
same way, however in our case we do not use `2^-128` precision for
seconds, we just keep integer number of seconds from the Unix epoch
start.