[TCFMM-8] Make set_position update ticks' liquidity
Problem: at the moment, updates for `liquidity_net` in ticks is missing
in `set_position`, while the article says it should be there.

Solution: along with updating `n_positions` field, update
`liquidity_net`.

Also I added a sanity check - ticks with non-zero `liquidity_net` should
never be garbage collected - that would point to a bug.

And slightly optimized the code - there is no necessity to call
`incr_n_positions` twice in each of the 4 `if` branches.