[TCFMM-11] Add `make clean`
Problem: it is convenient to have a way to clean everything and truly
rebuild the contract with `make`, without "nothing to be done" outcome.

Solution: add `make clean` that cleans `.tz` files and runs `stack
clean`.

I do not remove typescript files because they are included in git, not
sure if we should clean them and getting a non-clean git state.

Also I added all our unusual targets to `.PHONY` section.