[TCFMM-16] Fix `collect_fees` call
Problem: `collect_fees` expects the position to exist in the positions
big_map, otherwise it fails. However, `set_position` _always_ calls
`collect_fees`, even when initializing a new position that isn't in the
big_map yet.

Solution: do not call `collect_fees` when initializing a new position,
and set the fees to be collected to 0.