[TCFMM-12] Initialize ticks with "infinite" index at origination
Problem: Following the discussion in which the tick initialization is dependent
on existing "witnesses" ticks, there needs to be the two "extreme" ticks
(highest and lowest allowed indexes) at origination.

We need to pick a large number to represent this "infinite" index.

Solution: Add `default_storage` which is mean to be used at the
origination, use `const_max_tick` to represent the infinite indexes.