[TCFMM-25] Add `x_to_x_prime` entrypoint
Problem: This entrypoint allows a swap between two tokens that are
not directly paired in a pool, but are both paired with the same Y.
The intention behind it, is that ctez or a "base" token like it,
should become standard.

Solution: Add `x_to_x_prime` entrypoint, add a test to call it.