[TCFMM-14] Split test group
Problem: All the tests in `Test.SegCFMM` are grouped in a single test
group. This is unnecessary (we can simply prefix every test with
`test_`) and inconvenient (whenever a new test is added we have to add
it to the test gorup).

Solution: Prefix every test with `test_` and delete the test group.