[TCFMM-14] Fix sqrt_price of initial min/max ticks
Problem: `half_bps_pow` has a margin of error, and the `sqrt_prices` we
used in the default storage for the +/- inf ticks don't, this means we
might bump into funny scenarios.

There's an implicit invariant that, for ticks `i` and `j` , where `i <
j`:
> sqrt_price(i) <= sqrt_price(j)

However, when `i = -inf`  and `j = -inf+1`, then:
> sqrt_price(-inf) > sqrt_price(-inf + 1)

> 21 (hardcoded in the defaults) > 20 (calculated by half_bps_pow)

Solution: to account for the margin of error and avoid breaking this
invariant, we use `half_bps_pow` to calculate the +/- inf tick's
sqrt_prices.