[TCFMM-27] Refactor storage compilation and cleanup Makefile
Problem: storage compilation requires LIGO pragma because it is using
the 'main' as an 'ENTRY_POINT' for the 'compile-storage' LIGO command.
This however should be unnecessary because the storage is the same
regardless of the token types picked.

Solution: modify the modules and Makefile to use an identity contract
as the 'ENTRY_POINT' for the generation of the default storage.