[TCFMM-14] Fix garbage collection
Problem: In `set_position`, we use `cover_tick_with_position` to update
a tick's state and possibly garbage collect it if the position's being
deleted. However, later we need to retrieve the tick's state from the
bigmap again to calculate the fees the LP is owed. At this point, the
tick no longer exists, and the contract fails in line 194:

```
    (* Compute how much should be deposited / withdrawn to change liquidity by liquidity_net *)

    (* Grab cached prices for the interval *)
    let tick_u = get_tick ticks p.upper_tick_index internal_tick_not_exist_err in
    let tick_l = get_tick ticks p.lower_tick_index internal_tick_not_exist_err in
```

Solution: Move the garbage collection steps to the bottom of
`set_position`.