[TCFMM-14] Fix 'div by 0' bug in `update_timed_cumulatives`
Problem: When calculating the 'seconds per liquidity' accumulator at the beginning of each block, when the contract's liquidity is 0, the contract fails with a `div by 0` error.

Solution: when liquidity is 0, don't update the 'seconds per liquidity' accumulator.

It actually doesn't really matter how much we add to this accumulator when there is no active liquidity. When calculating a liquidity miner's rewards, we only care about the 'seconds per liquidity' accumulator's value _while_ the current tick was inside the position's range (i.e., _while_ the contract's liquidity was not zero).