[TCFMM-14] Add 'lower tick < upper_tick' check
Problem: At the moment, it's possible to set a position with a
lower_tick greater than the upper_tick. This breaks several invariants
of the contract, most notably: we end up with a range where the
liquidity is _negative_!

Solution: check that lower_tick is strictly lower than upper_tick.