[TCFMM-26] Add `get_position_info` entrypoint
Problem: in the liquidity mining contract I need a way to get the
current liquidity of a position when it is transferred to the mining
contract.

Solution: add `get_position_info` entrypoint for this.

When the new protocol is released, this (as well as other getters) must
be turned into an on-chain `view` for simplicity. (I hope LIGO will
also add support for embedded on-chain views very soon).

I also add `@layout:comb` for the entrypoint parameters with callback
because this way they start following the TZIP's A1 `View` pattern.