[TCFMM-14] Add 'Errors' haskell module
Problem: in the tests, we'll want to assert that some calls fail with a
certain error code. It would be best to avoid hardcoding these error
codes in the tests though.

Solution: Change the `generate_error_code.hs` script to also generate a
Haskell module with all the error codes.