[TCFMM-14] Fix bug in rounding of `dx`
Problem: When calculating how many tokens to give the user for a swap,
we're using `ceildiv` round the amount of tokens *up* to the nearest
integer.

However, rounding it up means we sometimes give the user ~1 extra token.
Over time, this effect can compound and slowly drain the contract's
balance faster than expected. This, in turn, means some Liquidity
Providers may not be able to liquidate their positions.

Solution: Round the amount of tokens to give the user *down*.