[TCFMM-26] Implement liquidity mining contract
Problem: one use case for Uniswap V3 oracles functionality is liquidity
mining, and we have to make a demo of this contract to ensure that all
the necessary API is exposed by our CFMM contract.

Solution: use the solidity Staker V3 implementation as example and
re-implement it in CamelLigo.

In some minor points I diverged from the implementation there, but the
overall design is the same.

I'm not sure I managed to grasp over their design, in some points it
remains controversial to me. So some `TODO` may be left.