[TCFMM-26] Update Makefile
Problem: we need a way to run the liquidity contract.

Solution: add it to makefile, add build of this contract to
`prepare_lib` target's dependencies so that we can easily run it in
tests later.

At this point I don't care about putting the contract to a separate
folder and separating Makefiles - one for CFMM, one for the liquidity
mining contract, and a root one - this is left for future work.