[TCFMM-26] Add documentation on the workflow of Staker
Problem: this liquidity mining contract (aka Staker contract) is not
documented, one has to guess how to work with it, collecting pieces of
information from the repo with the solidity implementation.

Solution: leave a brief description of the participating entities and of
the intended workflow.

As the contract is just demo, we don't yet bother with documenting
entrypoints and stuff, it is mostly for our internal usage in tests.