[TCFMM-37] Separate `liquidity_below_zero` errors
Problem: we use `internal_liquidity_below_zero` error to indicate that
position's liqudity went below zero. But this error is not internal, it
is quite a normal error.

Solution: we have two cases when this error is used - tick's liquidity
went below zero and position's liquidity went below zero. And these
errors are respectively an internal and a normal ones.

So we introduce two separate errors here instead of one.