[TCFMM-14] Fix bug when crossing `cur_tick_witness`
Problem:

1. Say we have have two adjacent positions at [-100, 100) and [-200,
   -100)
2. We place a x-to-y swap that exhausts the first position's liquidity
   and crosses the -100 tick
3. There's 1 X token left after crossing the -100 tick, so we attempt to
   execute another swap in the [-200, -100] range. In other words,
   `x_to_y_rec` recurses and we attempt a 2nd iteration.
4. A 1 token fee is applied, therefore there are no tokens left to swap
   in the [-200, -100] range. In other words, nothing happens in the 2nd
   iteration of `x_to_y_rec`.

In this scenario, the contract is left in a state that violates some
invariants. In step 3, when `x_to_y_rec` crossed the -100 tick, it:
* set `s.cur_tick_index` to -100
* it added the 2nd position's liquidity to the contract's global
  liquidity

This creates an inconsistency: The `cur_tick_index` is technically still
in the [-100, 100) range (remember, positions' ranges are half-open),
but the contract's liquidity is as if it was in the [-200, -100) range.

Solution: When `x_to_y_rec` recurses, we have to push the
`cur_tick_index` slightly further down, to, say, -101.