Allow negative values in range sums snapshot
Problem: following the formulas in the article, it is not necessary that
values like `seconds_inside` are exact (e.g. for `seconds_inside` - how
many seconds were spent in the given tick indices range).

When initializing `*_outside` values we have to do some
assumptions (e.g. that all seconds were spent below the current tick by
the moment of initialization), and thus if two ticks were initialized at
different moments, their `*_outside` values can be not in sync. And in
computing `*_inside` values we rely on tick's `*_outside` values.

The guarantee that we actually have: e.g. `seconds_outside` allows us
computing how many seconds were spent below or above some tick index,
but only with an error Δ, and this Δ remains constant while tick index
remains initialized. This holds for other `*_outside` values, and
`*_inside` values inherit this property.

Solution: clearify this in the spec and documentation, allow cumulative
values inside ranges to be negative.

I do not touch `fee_growth` value here because it is handled in a
parallel PR.