[TCFMM-14] Fix bug: overestimate amount of deposited Y tokens
Problem: we perform a `y_to_x` swap and we go into the recursive branch,
we calculate the `dy` to be deposited in the pool using
`Bitwise.shift_right`. `shift_right` truncates the amount of tokens,
rounding it down. This means that over time, the pool may end up with
fewer `y` tokens than expected, which in turns means the contract may
not have sufficient funds to liquidate all positions.

Solution: use `ceildiv` instead of `shift_right` to round `dy` *up*!