[TCFMM-14] Fix bug: in set_position, round delta.y up
Problem:

When rounding an amount of x/y tokens, the rule of thumb is to round the
number in the contract's favour. I.e., if we're depositing tokens in the
pool, we round it up, and if we're taking tokens from the pool, we round
it down.

The reasons for this are explained in `Note [Rounding the swap result]`.

Indeed, the comment in `set_position` states this:

> If I'm adding liquidity, x will be positive, I want to overestimate
> it, if x I'm taking away liquidity, I want to to underestimate what
> I'm receiving.

Using a concrete example:
* When liquidity_delta is positive and delta.y is 2.5 tokens, we should
  round it up to 3.
* When liquidity_delta is negative and delta.y is -2.5 tokens, we should
  round it up to -2.

However, the calculation of `delta.y` uses `shift_int`, which truncates
the number, that is, it rounds towards 0.
* When liquidity_delta is positive and delta.y is 2.5 tokens, shift_int
  rounds it down to 2.
* When liquidity_delta is negative and delta.y is -2.5 tokens, shift_int
  rounds it up to -2.

Solution: Instead of `shift_int k -80`, use `ceildiv_int k 2^80`.

`ceildiv_int` rounds towards +infinity, as opposed to `shift_int` which
rounds towards 0.