[TCFMM-14] Explain rounding behaviour
Problem: The contract has to round an amount of tokens up or down in
several places. The decision to round up vs down has big implications in
the contract's behaviour, and has an impact on the user as well.

However, they're poorly documented. There's only a few comments spread
throughout the contract's code, but they're not easily discoverable, and
the motivations behind those decisions are not made explicit.

Solution: Add a section to `specification.md` explaining where we round
up vs down and why.