[TCFMM-39] Include TZIP-16 metadata in contract storage
Problem: Currently we have a make target to produce a default storage
and one to produce TZIP-16 metadata, however:
- the target for the storage does not have an option to specify metadata
- the storage target should probably depend/use the metadata one directly
So, we should modify the $(OUT)/storage_default.tz to also generate and
include the metadata in the contract storage.

Solution: `$(OUT)/storage_default.tz` now generate the metadata and
includes it in the storage.