[TCFMM-14] Add invariant on liquidity and positions
Problem: one invariant we expect to hold is that, the current liquidity
is equal to the sum of liquidities of positions that cover the current
tick.

Maybe this invariant can be deduced from the other tested properties,
but for confidence, we better check this property directly and
explicitly.

Solution: check the invariant in tests.