[TCFMM-40] Impose granularity in tick initialization
Problem: There is the possibility that cost for swaps will get too
high when many nearby ticks are initialized, because we'll need to
traverse many of them. To prevent this from happening we should impose
a granularity to which ticks can be initialized.

In other words there should be a settable constant equivalent to the
`tickSpacing` in the whitepaper that establishes a number `n` so that
ticks can only be initialized if they are a multiple of `n`.

Solution: Add a constant for spacing between ticks, add tests to cover.