[TCFMM-41] Generalize `prepareSomeSegCFMM'`
Problem: most tests use very similar setup logic: create 2 fa2
contracts, create 1 cfmm contract, add the cfmm contract as an operator
for a given set of accounts.

Solution: We already have a couple of functions that handle this:
`prepareSomeSegCFMM` and `prepareSomeSegCFMM'`. If we generalize
`prepareSomeSegCFMM'` a bit more, then we can use this function is all
our tests. We ignore the `ProtocolFee` tests, those will be handled in
the next PR.