[TCFMM-13] Add arguments to some errors
Problem: for some errors it seems crucial to include error details.
For instance, in case deadline has been expired and the action has been
canceled, the user may want to get proofs of that deadline was violated
at the moment of execution.

Solution: for some user errors (with 1XX codes), include additional
info.