[TCFMM-42]  Perform balance requests in batches
Problem: The `balanceOf` and `balancesOf` helpers originate a consumer
contract every time they're called. If we need to retrieve the balance
of 2 accounts for 2 tokens, we'll execute 8 individual operations (2 * 2
contract originations + 2 * 2 calls).

Solution: First off, we need to separate the origination logic from the
calls. We need to be able to originate consumer contracts upfront, so
that we can then reuse them in multiple `balanceOf` calls.

Then, we add a `balancesOfMany` helper, that accepts `i` consumer
contracts and `j` accounts, and tries executes all calls (up to `i*j`
calls) in one single batch.

The `balanceOf` and `balancesOf` helpers are specializations of
`balancesOfMany`, and can be used:
* In emulator-only tests, where readability is more important than
  execution time.
* In network tests, when we're only interested in the balance of a
  single token / single account.