[TCFMM-42] Speed up `advanceSecs`
Problem: The `advanceSecs` needs to call some dummy entrypoint in the
contract to force the buffer to be updated. At the moment, it calls the
`Observe` entrypoint but, in order to do this, it needs to first
originate a consumer contract, which takes time when running on a
network.

Solution: Instead of calling the `Observe` entrypoint, we can call the
`Increase_observation_count` entrypoint with the argument `0`. This
should be a no-op.