[Chore] Remove extra-source-files section
Problem: In the package's `extra-source-files` section, we listed all our
contract *.tz files. This was done so that stack's `--file-watch`
reloads when those files change.

However, when we first build the segmented-cfmm library & app, those
files don't exist yet, so we get a lot of warnings.

They can't exist at that point, because we need the app to generate the
contract's metadata.

So the usual workflow is to:
* build the library & app
* build the ligo contract
* build the test suite, which uses the contract

Solution: Remove the contracts from the `extra-source-files` section.